Backend/requirements.txt
flask~=3.0.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Werkzeug==2.2.2
Flask-Cors==4.0.0
PyJWT~=2.8.0
requests~=2.25.1
SQLAlchemy~=2.0.22
Pillow~=10.0.1